fwdWindow

Extends an object representing a fish population for projecting into the future using the assumed equilibirum values used in the calculation of reference points

# S4 method for FLStock,FLBRP
fwdWindow(x, y, end = dims(x)$maxyear)

Arguments

x
The population object, for example of class FLStock

y
The reference points object, of class FLBRP

end
Final year of the extended object, always interpreted as a character

Value

An object of the same class a x, extended to year=end

Details

slts <- c("stock.wt", "landings.wt", "discards.wt", "catch.wt",
   "m", "mat", "harvest.spwn", "m.spwn")

Examples

data(ple4)

# Create the FLSR and FLBRP objects
psr <- fmle(as.FLSR(ple4, model="bevholt"), control=list(trace=0))
prp <- brp(FLBRP(ple4, sr=psr))

res <- fwdWindow(ple4, prp, end=2014)

#> Error in `[<-`(x = structure(c(0.038, 0.102, 0.157, 0.242, 0.325, 0.485, 0.719, 0.682, 0.844, 0.918, 0.041, 0.093, 0.18, 0.272, 0.303, 0.442, 0.577, 0.778, 0.793, 0.945, 0.045, 0.106, 0.173, 0.264, 0.329, 0.47, 0.65, 0.686, 0.908, 0.897, 0.038, 0.111, 0.181, 0.272, 0.364, 0.469, 0.633, 0.726, 0.845, 0.918, 0.037, 0.098, 0.185, 0.306, 0.337, 0.483, 0.579, 0.691, 0.779, 0.911, 0.036, 0.096, 0.173, 0.301, 0.424, 0.573, 0.684, 0.806, 0.873, 1.335, 0.041, 0.103, 0.176, 0.273, 0.378, 0.54, 0.663, 0.788, 0.882, 0.961, 0.024, 0.113, 0.184, 0.296, 0.373, 0.477, 0.645, 0.673, 0.845, 0.973, 0.031, 0.068, 0.198, 0.294, 0.333, 0.43, 0.516, 0.601, 0.722, 0.578, 0.031, 0.099, 0.127, 0.305, 0.403, 0.455, 0.503, 0.565, 0.581, 0.848, 0.029, 0.104, 0.179, 0.205, 0.442, 0.528, 0.585, 0.65, 0.703, 0.833, 0.055, 0.094, 0.175, 0.287, 0.344, 0.532, 0.592, 0.362, 0.667, 0.746, 0.047, 0.158, 0.188, 0.266, 0.344, 0.39, 0.565, 0.621, 0.679, 0.635, 0.043, 0.113, 0.236, 0.274, 0.369, 0.41, 0.468, 0.636, 0.732, 0.747, 0.051, 0.109, 0.251, 0.344, 0.413, 0.489, 0.512, 0.583, 0.696, 0.707, 0.056, 0.158, 0.218, 0.407, 0.473, 0.534, 0.579, 0.606, 0.655, 0.759, 0.037, 0.134, 0.237, 0.308, 0.468, 0.521, 0.566, 0.583, 0.617, 0.69, 0.049, 0.105, 0.217, 0.416, 0.437, 0.524, 0.57, 0.629, 0.652, 0.69, 0.063, 0.141, 0.187, 0.388, 0.483, 0.544, 0.61, 0.668, 0.704, 0.762, 0.082, 0.169, 0.226, 0.308, 0.484, 0.55, 0.593, 0.658, 0.694, 0.743, 0.064, 0.184, 0.265, 0.311, 0.405, 0.551, 0.627, 0.69, 0.667, 0.759, 0.064, 0.151, 0.319, 0.373, 0.411, 0.467, 0.547, 0.63, 0.704, 0.773, 0.062, 0.179, 0.258, 0.365, 0.414, 0.459, 0.543, 0.667, 0.764, 0.826, 0.049, 0.163, 0.289, 0.428, 0.444, 0.524, 0.582, 0.651, 0.778, 1.025, 0.041, 0.14, 0.239, 0.421, 0.473, 0.536, 0.57, 0.624, 0.707, 0.849, 0.048, 0.128, 0.25, 0.351, 0.49, 0.589, 0.631, 0.679, 0.726, 0.828, 0.045, 0.128, 0.242, 0.381, 0.494, 0.559, 0.624, 0.712, 0.754, 0.791, 0.048, 0.129, 0.216, 0.413, 0.464, 0.571, 0.649, 0.692, 0.787, 0.898, 0.048, 0.146, 0.232, 0.32, 0.452, 0.536, 0.635, 0.656, 0.764, 0.869, 0.043, 0.126, 0.245, 0.311, 0.44, 0.533, 0.692, 0.779, 0.888, 0.971, 0.036, 0.105, 0.2, 0.383, 0.401, 0.503, 0.573, 0.711, 0.747, 0.817, 0.036, 0.097, 0.172, 0.264, 0.426, 0.467, 0.547, 0.644, 0.706, 0.897, 0.039, 0.101, 0.192, 0.247, 0.362, 0.484, 0.553, 0.616, 0.759, 0.837, 0.043, 0.108, 0.176, 0.261, 0.343, 0.422, 0.555, 0.647, 0.701, 0.76, 0.048, 0.131, 0.184, 0.26, 0.342, 0.401, 0.463, 0.633, 0.652, 0.744, 0.043, 0.121, 0.199, 0.27, 0.318, 0.403, 0.5, 0.573, 0.683, 0.73, 0.05, 0.119, 0.208, 0.315, 0.33, 0.391, 0.49, 0.587, 0.633, 0.723, 0.053, 0.141, 0.214, 0.29, 0.36, 0.404, 0.462, 0.533, 0.653, 0.702, 0.05, 0.142, 0.254, 0.336, 0.399, 0.448, 0.509, 0.584, 0.678, 0.789, 0.044, 0.117, 0.229, 0.368, 0.39, 0.462, 0.488, 0.554, 0.66, 0.791, 0.035, 0.115, 0.233, 0.359, 0.439, 0.492, 0.521, 0.543, 0.627, 0.734, 0.038, 0.081, 0.207, 0.333, 0.474, 0.577, 0.581, 0.648, 0.656, 0.642, 0.044, 0.091, 0.15, 0.319, 0.437, 0.524, 0.586, 0.644, 0.664, 0.62, 0.051, 0.106, 0.165, 0.219, 0.408, 0.467, 0.649, 0.695, 0.656, 0.744, 0.061, 0.122, 0.202, 0.233, 0.331, 0.452, 0.56, 0.641, 0.798, 0.816, 0.048, 0.118, 0.213, 0.301, 0.319, 0.403, 0.446, 0.612, 0.685, 0.781, 0.057, 0.111, 0.227, 0.269, 0.344, 0.391, 0.464, 0.6, 0.714, 0.96, 0.047, 0.116, 0.201, 0.306, 0.384, 0.43, 0.489, 0.495, 0.78, 0.921, 0.053, 0.106, 0.216, 0.237, 0.378, 0.422, 0.434, 0.527, 0.621, 0.815, 0.052, 0.13, 0.19, 0.316, 0.354, 0.424, 0.439, 0.506, 0.583, 0.688, 0.047, 0.093, 0.235, 0.238, 0.337, 0.394, 0.458, 0.412, 0.526, 0.512, 0.048, 0.114, 0.196, 0.274, 0.355, 0.429, 0.484, 0.627, 0.598, 0.449, 0.052, 0.114, 0.194, 0.344, 0.373, 0.412, 0.472, 0.54, 0.565, 0.576, 0.053, 0.116, 0.179, 0.34, 0.361, 0.401, 0.448, 0.572, 0.568, 0.655, 0.039, 0.1, 0.187, 0.209, 0.355, 0.483, 0.438, 0.422, 0.53, 0.58, 0.052, 0.093, 0.142, 0.188, 0.331, 0.393, 0.484, 0.479, 0.48, 0.518, 0.043, 0.107, 0.153, 0.208, 0.32, 0.354, 0.434, 0.493, 0.662, 0.468, 0.048, 0.104, 0.158, 0.202, 0.312, 0.38, 0.439, 0.484, 0.458, 0.615), .Dim = c(10L, 1L, 1L, 1L, 1L, 58L), .Dimnames = list(    age = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10"    ), unit = "unique", season = "all", area = "unique", iter = "1",     year = c("1957", "1958", "1959", "1960", "1961", "1962",     "1963", "1964", "1965", "1966", "1967", "1968", "1969", "1970",     "1971", "1972", "1973", "1974", "1975", "1976", "1977", "1978",     "1979", "1980", "1981", "1982", "1983", "1984", "1985", "1986",     "1987", "1988", "1989", "1990", "1991", "1992", "1993", "1994",     "1995", "1996", "1997", "1998", "1999", "2000", "2001", "2002",     "2003", "2004", "2005", "2006", "2007", "2008", "2009", "2010",     "2011", "2012", "2013", "2014"))), i = 1:10, j = 1L, k = 1L,     l = 1L, m = 1L, n = 62:58, value = structure(c(0.0286666666666667,     0.0666666666666667, 0.123, 0.195333333333333, 0.269666666666667,     0.328333333333333, 0.372666666666667, 0.442333333333333,     0.473333333333333, 0.509333333333333), .Dim = c(10L, 1L,     1L, 1L, 1L, 1L))): subscript out of bounds