The FLR class for biological and economic reference points

Estimation of equilibirum reference points on the results obtained from fitting a population model can be carried out using the FLBRP class. Objects of this classs contain the information necessary for the calculation of population and fisheries quantities (abundance, catches, ...) under a range of levels of fishing mortality ('F'). From these, the values of those quantities that maximize or minimize the values of different indicators can be calculated.

# S4 method for missing,missing
FLBRP(
  ...,
  model = formula(rec ~ a),
  params = FLPar(1, params = "a"),
  fbar = FLQuant(seq(0, 4, length.out = 101), quant = "age")
)

Details

The most common input for the calculation of reference points is the result of an stock assessment model fit, usually provided as two objects of class FLStock and FLSR.

Slots

model

The formula for the stock-recruitment model, formula,.

params

Parameters of the stock-recruitment relationship, FLPar,.

refpts

The estimated reference points, FLPar,.

fbar

A vector of mean fishing mortality values for calculating the reference points, FLQuant,.

fbar.obs

The time series of observed mean fishing mortality, FLQuant,.

landings.obs

The time series of observed total landings, FLQuant,.

discards.obs

The time series of observed total discards, FLQuant,.

rec.obs

The time series of observed recruitment, FLQuant,.

ssb.obs

The time series of observed spawning stock biomasses, FLQuant,.

stock.obs

The time series of observed total stock abundance, FLQuant,.

profit.obs

The time series of profits, FLQuant,.

landings.sel

Selectivity at age for the landings, FLQuant,.

discards.sel

Selectivity at age for the discards, FLQuant,.

bycatch.harvest

X, Fishing mortality at age of bycatch species, FLQuant,.

stock.wt

Mean weight-at-age in the stock to use in calculations, FLQuant,.

landings.wt

Mean weight-at-age in the landings to use in calculations, FLQuant,.

discards.wt

Mean weight-at-age in the discards to use in calculations, FLQuant,.

bycatch.wt

Mean weight-at-age in the bycatch to use in calculations, FLQuant,.

m

Natural mortality at age, FLQuant,.

mat

Maturity at age, FLQuant,.

harvest.spwn

Proportion of harvest/fishing mortality before spawning, FLQuant,.

m.spwn

Proportion of natural mortality before spawning, FLQuant,.

availability

, FLQuant,.

price

Price at age, FLQuant,.

vcost

X, FLQuant,.

fcost

X, FLQuant,.

Validity

.obs slots dims 1:5

Slots containing observations, named *.obs, must share dimensions 1 to 5

by-age slots dims 1:5

Slots containing vectors of values at age, must share dimensions 1 to 5

You can inspect the class validity function by using getValidity(getClassDef('FLBRP'))

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing but not stored in any slot.

A standard method also exists for object of class FLStock, with or without additional information on the stock-recruits relationship. This can be provided as an object of class FLSR, one of class predictModel or as a list wqith elements named 'params', of class FLPar, and 'model', of class formula.

Methods

Methods exist for various calculations based on values stored in the class:

METHOD

Neque porro quisquam est qui dolorem ipsum.

See also

Author

The FLR Team

Examples

data(ple4)

# Create the FLSR and FLBRP objects
psr <- fmle(as.FLSR(ple4, model="bevholt"), control=list(trace=0))
prp <- brp(FLBRP(ple4, sr=psr))

summary(prp)

#> An object of class "FLBRP"
#> 
#> Name:  
#> Description:  
#> Quant: age 
#> Dims:  age 	year	unit	season	area	iter
#> 	10	101	1	1	1	1	
#> 
#> Range:  min	max	pgroup	minfbar	maxfbar 
#> 	1	10	10	2	6	
#> 
#> fbar          : [ 1 101 1 1 1 1 ], units =  f 
#> fbar.obs      : [ 1 61 1 1 1 1 ], units =  f 
#> landings.obs  : [ 1 61 1 1 1 1 ], units =  t 
#> discards.obs  : [ 1 61 1 1 1 1 ], units =  t 
#> rec.obs       : [ 1 61 1 1 1 1 ], units =  1000 
#> ssb.obs       : [ 1 61 1 1 1 1 ], units =  t 
#> stock.obs     : [ 1 61 1 1 1 1 ], units =  t 
#> profit.obs    : [ 1 61 1 1 1 1 ], units =  NA 
#> landings.sel  : [ 10 1 1 1 1 1 ], units =  NA 
#> discards.sel  : [ 10 1 1 1 1 1 ], units =  NA 
#> bycatch.harv  : [ 10 1 1 1 1 1 ], units =  f 
#> stock.wt      : [ 10 1 1 1 1 1 ], units =  kg 
#> landings.wt   : [ 10 1 1 1 1 1 ], units =  kg 
#> discards.wt   : [ 10 1 1 1 1 1 ], units =  kg 
#> bycatch.wt    : [ 10 1 1 1 1 1 ], units =  NA 
#> m             : [ 10 1 1 1 1 1 ], units =  m 
#> mat           : [ 10 1 1 1 1 1 ], units =   
#> harvest.spwn  : [ 10 1 1 1 1 1 ], units =   
#> m.spwn        : [ 10 1 1 1 1 1 ], units =   
#> availability  : [ 10 1 1 1 1 1 ], units =  NA 
#> price         : [ 10 1 1 1 1 1 ], units =  NA 
#> vcost         : [ 1 1 1 1 1 1 ], units =  NA 
#> fcost         : [ 1 1 1 1 1 1 ], units =  NA 
#> 
#> Model: 	rec ~ a * ssb/(b + ssb)
#>     params
#> iter       a     b
#>    1 1263310 93995
#> 
#> refpts:  calculated 

refpts(prp)

#> An object of class "FLPar"
#>         quant
#> refpt    harvest  yield    rec      ssb      biomass  revenue  cost    
#>   virgin 0.00e+00 0.00e+00 1.24e+06 4.22e+06 4.36e+06       NA       NA
#>   msy    1.88e-01 8.78e+04 1.17e+06 1.24e+06 1.35e+06       NA       NA
#>   crash  7.98e-01 1.88e-06 6.70e-05 4.98e-06 9.18e-06       NA       NA
#>   f0.1   1.55e-01 8.63e+04 1.19e+06 1.53e+06 1.65e+06       NA       NA
#>   fmax   2.08e-01 8.73e+04 1.16e+06 1.10e+06 1.21e+06       NA       NA
#>   spr.30 1.94e-01 8.77e+04 1.17e+06 1.20e+06 1.31e+06       NA       NA
#>   mey          NA       NA       NA       NA       NA       NA       NA
#>         quant
#> refpt    profit  
#>   virgin       NA
#>   msy          NA
#>   crash        NA
#>   f0.1         NA
#>   fmax         NA
#>   spr.30       NA
#>   mey          NA
#> units:  NA