fwdWindow

Extends an object representing a fish population for projecting into the future using the assumed equilibirum values used in the calculation of reference points

# S4 method for FLStock,FLBRP
fwdWindow(x, y, end = dims(x)$maxyear)

Arguments

x
The population object, for example of class FLStock

y
The reference points object, of class FLBRP

end
Final year of the extended object, always interpreted as a character

Value

An object of the same class a x, extended to year=end

Details

slts <- c("stock.wt", "landings.wt", "discards.wt", "catch.wt",
   "m", "mat", "harvest.spwn", "m.spwn")

Examples

data(ple4)

# Create the FLSR and FLBRP objects
psr <- fmle(as.FLSR(ple4, model="bevholt"), control=list(trace=0))
prp <- brp(FLBRP(ple4, sr=psr))

res <- fwdWindow(ple4, prp, end=dims(ple4)$maxyear + 10)